chore: update kafka and redpanda configuration
This commit updates the configuration files for Kafka and Redpanda.
Specifically, it changes the schema registry URLs to use the new service name
`redpanda.kafka.svc.cluster.local` instead of
`redpanda-0.redpanda.redpanda.svc.cluster.local`.
It also adds support for the
`process.env.PORT` environment variable in the `server.ts` file.
Additionally, it changes the namespace for the `dbz-connect` deployment and
`configure-kafka-connectors` job to `kafka`.
Finally, it adds a new `console` Helm chart for Redpanda and updates the `values.yaml`
file for Redpanda to disable the console by default.